feat: voice.go, lint fixes
